Refactor training and backtest pipeline
